+++ Src/App.fs
@@ -30,7 +30,9 @@ module App =
     [< EntryPoint >]
     [< STAThread >]
     let runEditorStandalone (args: string []) : int =
-        VelopackApp.Build().Run() //https://docs.velopack.io/getting-started/csharp
+        VelopackApp.Build()
+            .SetAutoApplyOnStartup(false)
+            .Run() //https://docs.velopack.io/getting-started/csharp
         let app  = Application() // do first so that pack Uris work
         let fesh = Initialize.everything (None, args)
         current <- fesh