+++ Src/Config/RecentlyUsedFiles.fs
@@ -19,7 +19,7 @@ type UsedFile = {
 
 type RecentlyUsedFiles  ( runContext:RunContext) =
 
-    let filePath0 = runContext.GetPathToSaveAppData("RecentlyUsedFiles.txt")
+    let filePath0 = runContext.GetPathToSaveAppData("Recently-Used-Files.txt")
 
     let writer = SaveReadWriter(filePath0,IFeshLog.printError)